refactor: rearange stuff
