+++ src/smsfusion/_transforms.py
@@ -1,6 +1,7 @@
 from warnings import warn
 
 import numpy as np
+from numba import njit
 from numpy.typing import NDArray
 
 
@@ -55,39 +56,142 @@ def _angular_matrix_from_euler(
     return t.reshape(-1, 3, 3)
 
 
-def _rot_matrix_from_euler(
-    alpha_beta_gamma: NDArray[np.float64],
-) -> NDArray[np.float64]:
+@njit
+def _rot_matrix_from_quaternion(q):
     """
-    Rotation matrix defined from Euler angles. The rotation matrix describes
-    rigid body rotation from-body-to-origin, according to xyz convention. That
-    is, first rotate about the x-axis (alpha - roll), then about the y-axis
-    (beta - pitch), and lastly about the z-axis (gamma - yaw).
+    Convert quaternion to rotation matrix.
+    """
+    q0, q1, q2, q3 = q
 
-    Rotating from-origin-to-body according to zyx convention, that is first
-    rotate about the z-axis (gamma - yaw), then about the y-axis
-    (beta - pitch), and lastly about the x-axis (alpha - roll) is achieved by
-    transposing (inverting) the rotation matrix defined here.
+    _2q1 = q1 + q1
+    _2q2 = q2 + q2
+    _2q3 = q3 + q3
+
+    _2q1q1 = q1 * _2q1
+    _2q1q2 = q1 * _2q2
+    _2q1q3 = q1 * _2q3
+    _2q2q2 = q2 * _2q2
+    _2q2q3 = q2 * _2q3
+    _2q3q3 = q3 * _2q3
+    _2q0q1 = q0 * _2q1
+    _2q0q2 = q0 * _2q2
+    _2q0q3 = q0 * _2q3
+
+    rot_00 = 1.0 - (_2q2q2 + _2q3q3)
+    rot_01 = _2q1q2 + _2q0q3
+    rot_02 = _2q1q3 - _2q0q2
+
+    rot_10 = _2q1q2 - _2q0q3
+    rot_11 = 1.0 - (_2q1q1 + _2q3q3)
+    rot_12 = _2q2q3 + _2q0q1
+
+    rot_20 = _2q1q3 + _2q0q2
+    rot_21 = _2q2q3 - _2q0q1
+    rot_22 = 1.0 - (_2q1q1 + _2q2q2)
+
+    rot = np.array(
+        [
+            [rot_00, rot_01, rot_02],
+            [rot_10, rot_11, rot_12],
+            [rot_20, rot_21, rot_22],
+        ]
+    )
+    return rot
+
+
+@njit
+def _euler_from_quaternion(q):
+    """
+    Convert quaternion to Euler angles (ZYX convention).
+    """
+    q0, q1, q2, q3 = q
+
+    _2q1 = q1 + q1
+    _2q2 = q2 + q2
+    _2q3 = q3 + q3
+
+    _2q1q1 = q1 * _2q1
+    _2q1q2 = q1 * _2q2
+    _2q1q3 = q1 * _2q3
+    _2q2q2 = q2 * _2q2
+    _2q2q3 = q2 * _2q3
+    _2q3q3 = q3 * _2q3
+    _2q0q1 = q0 * _2q1
+    _2q0q2 = q0 * _2q2
+    _2q0q3 = q0 * _2q3
+
+    rot_00 = 1.0 - (_2q2q2 + _2q3q3)
+    rot_01 = _2q1q2 + _2q0q3
+    rot_02 = _2q1q3 - _2q0q2
+
+    rot_12 = _2q2q3 + _2q0q1
+
+    rot_22 = 1.0 - (_2q1q1 + _2q2q2)
+
+    gamma = np.arctan2(rot_01, rot_00)
+    beta = -np.arcsin(rot_02)
+    alpha = np.arctan2(rot_12, rot_22)
+
+    return np.array([alpha, beta, gamma])
+
+
+@njit
+def _gamma_from_quaternion(q):
+    """
+    Get yaw from quaternion (ZYX convention).
+    """
+    q0, q1, q2, q3 = q
+
+    _2q2 = q2 + q2
+    _2q3 = q3 + q3
+
+    _2q1q2 = q1 * _2q2
+    _2q2q2 = q2 * _2q2
+    _2q3q3 = q3 * _2q3
+    _2q0q3 = q0 * _2q3
+
+    rot_00 = 1.0 - (_2q2q2 + _2q3q3)
+    rot_01 = _2q1q2 + _2q0q3
+
+    yaw = np.arctan2(rot_01, rot_00)
+    return yaw
+
+
+@njit
+def _angular_matrix_from_quaternion(q):
+    """
+    Angular transformation matrix, such that dq/dt = T(q) * omega.
+    """
+    return 0.5 * np.array(
+        [
+            [-q[1], -q[2], -q[3]],
+            [q[0], -q[3], q[2]],
+            [q[3], q[0], -q[1]],
+            [-q[2], q[1], q[0]],
+        ]
+    )
+
+
+@njit
+def _rot_matrix_from_euler(euler):
+    """
+    Rotation matrix defined from Euler angles (ZYX convention). Note that the rotation
+    matrix describes the rigid body rotation from-origin-to-body, according to ZYX convention.
+
 
     Parameters
     ----------
-    alpha_beta_gamma : ndarray
-        Euler angle about x-axis (alpha-roll), y-axis (beta-pitch), and z-axis
-        (gamma-yaw) in radians.
+    euler : 1D array (3,)
+        Euler angle in radians given as (roll, pitch, yaw) but rotaions are applied
+        according to the ZYX convention. That is, **yaw -> pitch -> roll**.
 
     Return
     ------
-    rot : array (Nx3x3)
-        3D rotation matrix.
-
-    Notes
-    -----
-    The shape of 'alpha_beta_gamma' determines the casting rule for the output.
-    If 'alpha_beta_gamma' is ndarray (Nx3), then the output is (Nx3x3).
+    rot : ndarray (3, 3)
+        Rotation matrix.
 
     """
-    alpha, beta, gamma = alpha_beta_gamma.T
-
+    alpha, beta, gamma = euler
     cos_gamma = np.cos(gamma)
     sin_gamma = np.sin(gamma)
     cos_beta = np.cos(beta)
@@ -95,19 +199,19 @@ def _rot_matrix_from_euler(
     cos_alpha = np.cos(alpha)
     sin_alpha = np.sin(alpha)
 
+    rot_01 = cos_beta * sin_gamma
     rot_00 = cos_gamma * cos_beta
-    rot_01 = cos_gamma * sin_beta * sin_alpha - cos_alpha * sin_gamma
-    rot_02 = cos_gamma * cos_alpha * sin_beta + sin_gamma * sin_alpha
+    rot_02 = -sin_beta
 
-    rot_10 = cos_beta * sin_gamma
+    rot_10 = cos_gamma * sin_beta * sin_alpha - cos_alpha * sin_gamma
     rot_11 = cos_gamma * cos_alpha + sin_gamma * sin_beta * sin_alpha
-    rot_12 = cos_alpha * sin_gamma * sin_beta - cos_gamma * sin_alpha
+    rot_12 = cos_beta * sin_alpha
 
-    rot_20 = -sin_beta
-    rot_21 = cos_beta * sin_alpha
+    rot_20 = cos_gamma * cos_alpha * sin_beta + sin_gamma * sin_alpha
+    rot_21 = cos_alpha * sin_gamma * sin_beta - cos_gamma * sin_alpha
     rot_22 = cos_beta * cos_alpha
 
     rot = np.array(
-        [[rot_00, rot_10, rot_20], [rot_01, rot_11, rot_21], [rot_02, rot_12, rot_22]]
-    ).T
-    return rot.reshape(-1, 3, 3)
+        [[rot_00, rot_01, rot_02], [rot_10, rot_11, rot_12], [rot_20, rot_21, rot_22]]
+    )
+    return rot
