+++ tests/test_transforms.py
@@ -1,4 +1,16 @@
+"""
+IMPORTANT
+---------
+
+SciPy Rotation implementation is used as reference in tests. However, SciPy
+operates with active rotations, whereas passive rotations are considered here. Keep in
+mind that passive rotations is simply the inverse active rotations and vice versa.
+"""
+
+
 import numpy as np
+import pytest
+from scipy.spatial.transform import Rotation
 
 from smsfusion import _transforms
 
@@ -123,121 +135,103 @@ class Test__angular_matrix_from_euler:
         )
 
 
-class Test__rot_matrix_from_euler:
+@pytest.mark.parametrize(
+    "q",
+    [
+        np.array([0.96591925, -0.25882081, 0.0, 0.0], dtype=float),  # about x-axis
+        np.array([0.96591925, 0.0, -0.25882081, 0.0], dtype=float),  # about y-axis
+        np.array([0.96591925, 0.0, 0.0, -0.25882081], dtype=float),  # about z-axis
+    ],
+)
+def test_rot_matrix_from_quaternion(q):
+    rot_matrix = _transforms._rot_matrix_from_quaternion(q)
+    rot_matrix_expect = Rotation.from_quat(q[[1, 2, 3, 0]]).inv().as_matrix()
+    np.testing.assert_array_almost_equal(rot_matrix, rot_matrix_expect, decimal=3)
+
+
+@pytest.mark.parametrize(
+    "q",
+    [
+        np.array([0.96591925, 0.0, 0.0, -0.25882081], dtype=float),  # pure about z-axis
+        np.array([0.93511153, 0.0, 0.25056579, -0.25056579], dtype=float),  # mixed
+    ],
+)
+def test__gamma_from_quaternion(q):
+    gamma = _transforms._gamma_from_quaternion(q)
+    gamma_expect = Rotation.from_quat(q[[1, 2, 3, 0]]).as_euler("ZYX", degrees=False)[0]
+    np.testing.assert_almost_equal(gamma, gamma_expect, decimal=3)
+
+
+def test__angular_matrix_from_quaternion():
+    q = np.array([1.0, 2.0, 3.0, 4.0]) / np.sqrt(30.0)
+
+    T_expect = (
+        np.array(
+            [
+                [-q[1], -q[2], -q[3]],
+                [q[0], -q[3], q[2]],
+                [q[3], q[0], -q[1]],
+                [-q[2], q[1], q[0]],
+            ]
+        )
+        / 2.0
+    )
+
+    T = _transforms._angular_matrix_from_quaternion(q)
+
+    np.testing.assert_almost_equal(T, T_expect)
+
+
+@pytest.mark.parametrize(
+    "angle, axis, euler",
+    [
+        (
+            np.radians(10.0),
+            np.array([0.0, 0.0, 1.0]),
+            np.array([0.0, 0.0, np.radians(10.0)]),
+        ),  # pure yaw
+        (
+            np.radians(10.0),
+            np.array([0.0, 1.0, 0.0]),
+            np.array([0.0, np.radians(10.0), 0.0]),
+        ),  # pure pitch
+        (
+            np.radians(10.0),
+            np.array([1.0, 0.0, 0.0]),
+            np.array([np.radians(10.0), 0.0, 0.0]),
+        ),  # pure roll
+        (
+            np.radians(10.0),
+            np.array([1.0 / np.sqrt(3.0), 1.0 / np.sqrt(3.0), 1.0 / np.sqrt(3.0)]),
+            np.array([0.1059987325729154, 0.0953360919950474, 0.1059987325729154]),
+        ),  # mixed
+    ],
+)
+def test__euler_from_quaternion(angle, axis, euler):
+    q = np.array(
+        [
+            np.cos(angle / 2),
+            np.sin(angle / 2) * axis[0],
+            np.sin(angle / 2) * axis[1],
+            np.sin(angle / 2) * axis[2],
+        ]
+    )
+
+    alpha_beta_gamma = _transforms._euler_from_quaternion(q)
+    np.testing.assert_array_almost_equal(alpha_beta_gamma, euler, decimal=16)
+
+
+@pytest.mark.parametrize("euler", [
+    np.array([10., 0.0, 0.0]),  # pure roll
+    np.array([0.0, 10.0, 0.0]),  # pure pitch
+    np.array([0.0, 0.0, 10.0]),  # pure yaw
+    np.array([10.0, -10.0, 10.0]),  # mixed
+])
+def test__rot_matrix_from_euler(euler):
     """
-    Reference cases are obtained from https://www.andre-gaschler.com/rotationconverter/
-    and https://github.com/gaschler/rotationconverter.
-
-    Note that Euler angles ZYX is selected, which corresponds to the rotation
-    matrix defined in this package. (ZYX active rotaiton -> XYZ passive rotation).
+    The Numba optimized implementaiton uses from-origin-to-body (zyx) convention,
+    where also the resulting rotation matrix is from-origin-to-body.
     """
-
-    def test_pure_roll(self):
-        rot_matrix = _transforms._rot_matrix_from_euler(np.radians((30.0, 0.0, 0.0)))
-
-        rot_matrix_expected = np.array(
-            [[1.000, 0.000, 0.000], [0.000, 0.866, -0.500], [0.000, 0.500, 0.866]]
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[0], rot_matrix_expected, decimal=3
-        )
-
-    def test_pure_pitch(self):
-        rot_matrix = _transforms._rot_matrix_from_euler(np.radians((0.0, 30.0, 0.0)))
-
-        rot_matrix_expected = np.array(
-            [[0.866, 0.000, 0.500], [0.000, 1.000, 0.000], [-0.500, 0.000, 0.866]]
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[0], rot_matrix_expected, decimal=3
-        )
-
-    def test_pure_yaw(self):
-        rot_matrix = _transforms._rot_matrix_from_euler(np.radians((0.0, 0.0, 30.0)))
-
-        rot_matrix_expected = np.array(
-            [[0.866, -0.500, 0.000], [0.500, 0.866, 0.000], [0.000, 0.000, 1.000]]
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[0], rot_matrix_expected, decimal=3
-        )
-
-    def test_references(self):
-        # case 1
-        rot_matrix = _transforms._rot_matrix_from_euler(np.radians((30.0, 15.0, 20.0)))
-
-        rot_matrix_expected = np.array(
-            [
-                [0.9076734, -0.1745930, 0.3816364],
-                [0.3303661, 0.8580583, -0.3931846],
-                [-0.2588190, 0.4829629, 0.8365163],
-            ]
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[0], rot_matrix_expected, decimal=3
-        )
-
-        # case 2
-        rot_matrix = _transforms._rot_matrix_from_euler(np.radians((15.0, 45.0, 5.0)))
-
-        rot_matrix_expected = np.array(
-            [
-                [0.7044160, 0.0981303, 0.7029712],
-                [0.0616284, 0.9782008, -0.1983057],
-                [-0.7071068, 0.1830127, 0.6830127],
-            ]
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[0], rot_matrix_expected, decimal=3
-        )
-
-    def test_vectorized(self):
-        euler_vector = np.array(
-            [
-                np.radians((30.0, 15.0, 20.0)),
-                np.radians((15.0, 45.0, 5.0)),
-                np.radians((25.0, 125.0, -35.0)),
-            ]
-        )
-        rot_matrix = _transforms._rot_matrix_from_euler(euler_vector)
-
-        rot_matrix_expected_0 = np.array(
-            [
-                [0.9076734, -0.1745930, 0.3816364],
-                [0.3303661, 0.8580583, -0.3931846],
-                [-0.2588190, 0.4829629, 0.8365163],
-            ]
-        )
-
-        rot_matrix_expected_1 = np.array(
-            [
-                [0.7044160, 0.0981303, 0.7029712],
-                [0.0616284, 0.9782008, -0.1983057],
-                [-0.7071068, 0.1830127, 0.6830127],
-            ]
-        )
-
-        rot_matrix_expected_2 = np.array(
-            [
-                [-0.4698463, 0.8034179, 0.3657378],
-                [0.3289899, 0.5438383, -0.7720140],
-                [-0.8191521, -0.2424039, -0.5198368],
-            ]
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[0], rot_matrix_expected_0, decimal=3
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[1], rot_matrix_expected_1, decimal=3
-        )
-
-        np.testing.assert_array_almost_equal(
-            rot_matrix[2], rot_matrix_expected_2, decimal=3
-        )
+    out = _transforms._rot_matrix_from_euler(euler)
+    expected = Rotation.from_euler("ZYX", euler[::-1]).inv().as_matrix()
+    np.testing.assert_array_almost_equal(out, expected)
